Moved pytest fixtures to conftest.py
